+++ src/usr/ld.a/lib_stat/st_stdint.h
@@ -20,25 +20,25 @@
 #define ST_STDINT_H
 
 typedef unsigned long long st_uint64_t;
-typedef unsigned int st_uint32_t;
+typedef unsigned long st_uint32_t;
 typedef unsigned short st_uint16_t;
 typedef unsigned char st_uint8_t;
 
 typedef long long  st_int64_t;
-typedef int st_int32_t;
+typedef long st_int32_t;
 typedef short st_int16_t;
 typedef char st_int8_t;
 
 // size 
-typedef unsigned int st_size_t;
+typedef unsigned long st_size_t;
 
 // Signed pointer-sized integer 
 typedef long st_intptr_t;
 // Unsigned pointer-sized integer
 typedef unsigned long st_uintptr_t;
 
-typedef unsigned int st_uint_t;
-typedef int st_int_t;
+typedef unsigned long st_uint_t;
+typedef long st_int_t;
 
 
 #endif
